requirements.txt
ipdb
Flask==1.1.2
python-dotenv==0.17.0
requests==2.25.1
google-api-core==1.26.3
google-auth==1.28.1
google-cloud-pubsub==2.4.1